+++ Cosette/Engine/Ai/Search/SearchConstants.cs
@@ -11,7 +11,7 @@
         public const int DefaultPawnHashTableSize = 8;
 
         public const int NullWindowMinimalDepth = 5;
-        public const int NullWindowDepthReduction = 3;
+        public const byte NullWindowDepthReduction = 3;
 
         public const int LMRMinimalDepth = 2;
         public const int LMRMovesWithoutReduction = 2;
